fix: don't filter single runs in the comparison view
This fixes an issue where single search runs were being filtered out of
the comparison view in runs view. We were erroneously applying a filter
to the search results.
